requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
requests>=2.27.1
websocket-client>=1.3.1
# kraken-wsclient-py
celery
redis
